requirements.txt
requests
python-dotenv
matplotlib
flask
